Consume provider-free evaluation kit
