+++ Sources/AFMKit/AFMEvaluationExports.swift
@@ -1,0 +1,3 @@
+// Preserve the existing `import AFMKit` evaluation surface while keeping the
+// host-owned suite discovery and bundled resources independently buildable.
+@_exported import AFMEvaluationHost
